requirements.txt
openai==1.35.10
python-dotenv==1.0.1
httpx==0.25.2
